Refactor registrar into multiple files
