+++ src/cli/register.py
@@ -37,7 +37,7 @@ def register_dataset(args):
     )
 
     # Register new dataset.
-    new_id = datareg.Registrar.register_dataset(
+    new_id = datareg.Registrar.dataset.create(
         args.relative_path,
         args.version,
         name=args.name,
