+++ src/dataregistry/__init__.py
@@ -1,7 +1,7 @@
 from ._version import __version__
 from .db_basic import *
 from .registrar import *
-from .registrar_util import *
+#from .registrar_util import *
 from .query import *
 from .git_util import *
 from .DataRegistry import DataRegistry
