+++ src/dataregistry/registrar/__init__.py
@@ -1,0 +1,1 @@
+from .registrar import _OWNER_TYPES, Registrar
